# docs/homework_openapi.yaml
components:
  schemas:
...
    ErrorResponse:
      properties:
        message:
          example: 操作失败
          type: string
        success:
          example: false
          type: boolean
      type: object
    Year:
      properties:
        created_at:
          example: '2024-01-01 10:00:00'
          format: date-time
          type: string
        id:
          example: 1
          type: integer
        is_active:
          example: true
          type: boolean
        name:
          example: 2024学年度
          type: string
        teacher_id:
          example: 10
          type: integer
        teacher_name:
          example: 张老师
          type: string
        year:
          example: 2024
          type: integer
      type: object
...
    SuccessResponse:
      properties:
        data:
          type: object
        message:
          example: 操作成功
          type: string
        success:
          example: true
          type: boolean
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: 'JWT Token认证，需要在请求头中携带 Authorization: Bearer <token>'
      scheme: bearer
      type: http
info:
  description: 作业库系统后端接口文档，包含学生端和教师端的所有作业相关功能
  title: 作业库系统API
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /api/v2/years:
    get:
      description: 获取年份列表（教师可以查看自己创建的年份，管理员可以查看所有年份）
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/Year'
                      type: array
                  type: object
          description: 获取成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 获取年份列表
      tags:
      - 年份管理
    post:
      description: 创建年份（仅教师和管理员）
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  description: 年份名称（如：2024学年度）
                  example: 2024学年度
                  type: string
                year:
                  description: 年份（整数）
                  example: 2024
                  format: int32
                  type: integer
              required:
              - year
              - name
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Schema'
                - properties:
                    data:
                      $ref: '#/components/schemas/Year'
                  type: object
                type: object
          description: 创建成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 年份已存在
      security:
      - bearerAuth: []
      summary: 创建年份
      tags:
      - 年份管理
  /api/v2/years/{year_id}:
    delete:
      description: 删除年份（软删除）（仅创建者和管理员）
      parameters:
      - description: 年份ID
        in: path
        name: year_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: 删除成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 年份不存在
      security:
      - bearerAuth: []
      summary: 删除年份
      tags:
      - 年份管理
    put:
      description: 更新年份（仅创建者和管理员）
      parameters:
      - description: 年份ID
        in: path
        name: year_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                is_active:
                  description: 是否激活
                  example: true
                  type: boolean
                name:
                  description: 年份名称
                  example: 2024学年度
                  type: string
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      $ref: '#/components/schemas/Year'
                  type: object
                type: object
          description: 更新成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 年份不存在
      security:
      - bearerAuth: []
      summary: 更新年份
      tags:
      - 年份管理
servers:
- description: Local development server
  url: http://localhost:5000
- description: Local network server
  url: http://192.168.8.190:5000
...